+++ src/compiler/parse_global.c
@@ -421,7 +421,7 @@ static inline TypeInfo *parse_base_type(ParseContext *c)
 		CONSUME_OR_RET(is_lparen ? TOKEN_RPAREN : TOKEN_RBRACKET, poisoned_type_info);
 		RANGE_EXTEND_PREV(type_info);
 		// TODO remove in 0.7
-		if (is_lparen && !compiler.context.silence_deprecation)
+		if (is_lparen && !compiler.build.silence_deprecation)
 		{
 			SEMA_NOTE(type_info, "'$vatype(...)' is deprecated, use '$vatype[...]' instead.");
 		}